Move ephemeral search helpers into ephemeral store
